Add I2C API
This is only a first iteration that has quite severe limitations. Please
refer to the API documentation (part of this commit) for details.
